+++ src/components/Grid/Column.js
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import omit from 'lodash.omit';
 
-import { flavorable, variationable } from '../../decorators';
-
 const validBreakpoints = [
   'small',
   'medium',
@@ -18,12 +16,17 @@ const breakPointProp = breakpoint => `${breakpoint}-sizeOf`;
 const sizeRegex = /^([1-9]|1[0-2])-([1-9]|1[0-2])$/;
 
 export const Column = (props) => {
-  const { align, children, className, omitCol, ...rest } = props;
+  const { align, children, className, flavor, variation, omitCol, ...rest } = props;
+
+  const flavorClasses = Array.isArray(flavor) ? flavor.map(f => `slds-col_${f}`) : `slds-col_${flavor}`;
+  const variationClasses = Array.isArray(variation) ? variation.map(f => `slds-${f}`) : `slds-${variation}`;
 
   const sldsClasses = [
     { 'slds-col': !omitCol },
     { [`slds-align-${align}`]: !!align },
-    className
+    className,
+    flavorClasses,
+    variationClasses,
   ];
 
   const breakpoints = Array.from(validBreakpoints);
@@ -47,30 +50,6 @@ export const Column = (props) => {
     <div {...restProps} className={cx(sldsClasses)}>{children}</div>
   );
 };
-
-Column.flavors = [
-  'bump-left',
-  'bump-right',
-  'bump-top',
-  'bump-bottom',
-  'padded',
-  'padded-large',
-  'padded-small',
-  'rule-right',
-  'rule-left',
-  'rule-top',
-  'rule-bottom',
-];
-
-Column.variations = [
-  'has-flexi-truncate',
-  'no-flex',
-  'no-space',
-  'grow',
-  'grow-none',
-  'shrink',
-  'shrink-none',
-];
 
 const sizeOfPropType = (props, propName) => {
   const size = props[propName];
@@ -98,6 +77,8 @@ const sizeOfPropType = (props, propName) => {
 Column.defaultProps = {
   align: 'top',
   className: null,
+  flavor: [],
+  variation: [],
   omitCol: false,
 };
 
@@ -115,6 +96,54 @@ Column.propTypes = {
    */
   className: PropTypes.string,
   /**
+   * flavor
+   */
+  flavor: PropTypes.oneOfType([PropTypes.oneOf([
+    'bump-left',
+    'bump-right',
+    'bump-top',
+    'bump-bottom',
+    'padded',
+    'padded-large',
+    'padded-small',
+    'rule-right',
+    'rule-left',
+    'rule-top',
+    'rule-bottom',
+  ]), PropTypes.arrayOf(PropTypes.oneOf([
+    'bump-left',
+    'bump-right',
+    'bump-top',
+    'bump-bottom',
+    'padded',
+    'padded-large',
+    'padded-small',
+    'rule-right',
+    'rule-left',
+    'rule-top',
+    'rule-bottom',
+  ]))]),
+  /*
+   * variation
+   */
+  variation: PropTypes.oneOfType([PropTypes.oneOf([
+    'has-flexi-truncate',
+    'no-flex',
+    'no-space',
+    'grow',
+    'grow-none',
+    'shrink',
+    'shrink-none',
+  ]), PropTypes.arrayOf(PropTypes.oneOf([
+    'has-flexi-truncate',
+    'no-flex',
+    'no-space',
+    'grow',
+    'grow-none',
+    'shrink',
+    'shrink-none',
+  ]))]),
+  /**
    * non-responsive sizeOf
    */
   sizeOf: sizeOfPropType, // eslint-disable-line react/require-default-props
@@ -129,6 +158,4 @@ Column.propTypes = {
   }, {}),
 };
 
-export default variationable(
-  flavorable(Column, 'col')
-);
+export default Column;