+++ src/components/Grid/__tests__/Column.spec.js
@@ -34,4 +34,20 @@ describe('<Column />', () => {
     mounted.setProps({ omitCol: true });
     expect(mounted.find('.slds-col').length).toBe(0);
   });
+
+  it('applies flavoring', () => {
+    mounted.setProps({ flavor: 'padded' });
+    expect(mounted.find('.slds-col').hasClass('slds-col_padded')).toBeTruthy();
+    mounted.setProps({ flavor: ['padded', 'bump-right'] });
+    expect(mounted.find('.slds-col').hasClass('slds-col_padded')).toBeTruthy();
+    expect(mounted.find('.slds-col').hasClass('slds-col_bump-right')).toBeTruthy();
+  });
+
+  it('applies variationing', () => {
+    mounted.setProps({ variation: 'shrink' });
+    expect(mounted.find('.slds-col').hasClass('slds-shrink')).toBeTruthy();
+    mounted.setProps({ variation: ['shrink', 'no-flex'] });
+    expect(mounted.find('.slds-col').hasClass('slds-shrink')).toBeTruthy();
+    expect(mounted.find('.slds-col').hasClass('slds-no-flex')).toBeTruthy();
+  });
 });