+++ src/components/Grid/__tests__/Container.spec.js
@@ -20,4 +20,11 @@ describe('<Container />', () => {
     expect(mounted.find('div').hasClass('foo')).toBeTruthy();
     expect(mounted.find('div').prop('data-test')).toEqual('bar');
   });
+
+  it('applies flavoring', () => {
+    mounted.setProps({ flavor: 'small' });
+    expect(mounted.find('.slds-container_small').length).toEqual(1);
+    mounted.setProps({ flavor: ['center', 'small'] });
+    expect(mounted.find('.slds-container_small').hasClass('slds-container_center')).toBeTruthy();
+  });
 });