+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 # Unreleased
 
-* Breaking: `Table` flavor and variation usage
+* Breaking: `Grid` flavor & variation usage
+* Breaking: `Column` flavor & variation usage
+* Breaking: `Container` flavor & variation usage
+* Breaking: `Table` flavor & variation usage
 * Breaking: `Popover` theme usage, renamed customLayout to customHeaderTheme
 * Breaking: `StatefulButton` flavor usage
 * Breaking: `Button` flavor usage